# Refactor and speed up tests