--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -2,11 +2,11 @@ bumpversion>=0.5.3,<1
 flake8==3.5.0
 mypy==0.600
 hypothesis==3.44.26
+pytest==3.6.0
 pytest-asyncio==0.8.0
 pytest-cov==2.5.1
-pytest-logging>=2015.11.4
-pytest-xdist==1.18.1
-pytest==3.1.2
-tox==2.7.0
+pytest-logging>=0.3.0
+pytest-xdist==1.22.2
+tox==3.0.0
 eth-tester[py-evm]==0.1.0-beta.26
 git+git://github.com/ethereum/vyper.git@08ba8ed7c3c84d44edda85ff28c96bd1e2d867fe