--- a/tests/contract/test_registry_management.py
+++ b/tests/contract/test_registry_management.py
@@ -118,7 +118,7 @@ def test_normal_deregister(smc_handler):  # noqa: F811
     assert notary_pool_length == 1
 
     # Fast foward
-    mine(w3, 20 - w3.eth.blockNumber)
+    fast_forward(smc_handler, 1)
 
     # Deregister notary 0
     smc_handler.deregister_notary(private_key=notary_0.private_key)
@@ -148,7 +148,7 @@ def test_deregister_then_register(smc_handler):  # noqa: F811
     assert notary_pool_length == 1
 
     # Fast foward
-    mine(w3, 20 - w3.eth.blockNumber)
+    fast_forward(smc_handler, 1)
 
     # Deregister notary 0
     smc_handler.deregister_notary(private_key=notary_0.private_key)
@@ -187,7 +187,7 @@ def test_normal_release_notary(smc_handler):  # noqa: F811
     assert notary_pool_length == 1
 
     # Fast foward
-    mine(w3, 20 - w3.eth.blockNumber)
+    fast_forward(smc_handler, 1)
 
     # Deregister notary 0
     smc_handler.deregister_notary(private_key=notary_0.private_key)
@@ -219,7 +219,7 @@ def test_instant_release_notary(smc_handler):  # noqa: F811
     assert notary_pool_length == 1
 
     # Fast foward
-    mine(w3, 20 - w3.eth.blockNumber)
+    fast_forward(smc_handler, 1)
 
     # Deregister notary 0
     smc_handler.deregister_notary(private_key=notary_0.private_key)
@@ -260,7 +260,7 @@ def test_deregister_and_new_notary_register(smc_handler):  # noqa: F811
     assert empty_slots_stack_top == 0
 
     # Fast foward
-    mine(w3, 20 - w3.eth.blockNumber)
+    fast_forward(smc_handler, 1)
 
     # Deregister notary 2
     smc_handler.deregister_notary(private_key=notary_2.private_key)