--- a/tests/contract/test_submit_vote.py
+++ b/tests/contract/test_submit_vote.py
@@ -3,9 +3,7 @@ import pytest
 from sharding.handler.utils.web3_utils import (
     mine,
 )
-from tests.handler.fixtures import (  # noqa: F401
-    smc_handler,
-)
+
 from tests.contract.utils.common_utils import (
     batch_register,
     fast_forward,