### What was wrong?
Investigating #92

### How was it fixed?
1. Updated `pytest` to `3.6.0`
2. It seems `smc_handler` fixture is being used everywhere. Added `tests/conftest.py` for it.
3. Set testing `NOTARY_LOCKUP_LENGTH` from `60` to `30`.
4. Modified the `block_number` in `mine(block_number)` from `50` or `100` to `20`.

#### Benchmark
##### 1. previous code + `pytest==3.1.2`
```
===================================================== slowest 5 test durations ======================================================
40.59s call     tests/contract/test_registry_management.py::test_normal_release_notary
36.59s call     tests/handler/test_shard_tracker.py::test_status_checking_functions
31.91s call     tests/contract/test_log_emission.py::test_log_emission
5.12s call     tests/contract/test_submit_vote.py::test_normal_submit_vote
4.12s call     tests/contract/test_submit_vote.py::test_submit_vote_by_notary_sampled_multiple_times
==================================================== 56 passed in 168.27 seconds =====================================================
```

##### 2. previous code + `pytest==3.6.0`
```
===================================================== slowest 5 test durations ======================================================
35.01s call     tests/handler/test_shard_tracker.py::test_status_checking_functions
31.85s call     tests/contract/test_log_emission.py::test_log_emission
29.25s call     tests/contract/test_registry_management.py::test_normal_release_notary
2.64s call     tests/contract/test_notary_sample.py::test_get_sample_result
2.56s call     tests/contract/test_submit_vote.py::test_normal_submit_vote
==================================================== 56 passed in 134.44 seconds =====================================================
```

##### 3. new code + `pytest==3.6.0`
```
===================================================== slowest 5 test durations ======================================================
13.90s call     tests/handler/test_shard_tracker.py::test_status_checking_functions
12.63s call     tests/contract/test_registry_management.py::test_normal_release_notary
12.32s call     tests/contract/test_log_emission.py::test_log_emission
3.21s call     tests/contract/test_notary_sample.py::test_get_sample_result
2.62s call     tests/contract/test_submit_vote.py::test_normal_submit_vote
===================================================== 56 passed in 75.93 seconds =====================================================
```

#### Cute Animal Picture

![300356600_d22d497ba0_z](https://user-images.githubusercontent.com/9263930/40589094-1016616c-621a-11e8-9cc1-9ef3a86d867a.jpg)